GulpSimulation cleanup and test coverage

+++ tests/tests_integration/task/gulp/GulpSimulation/test_GulpSimulation.py
@@ -2,6 +2,7 @@ import pytest
 import os, shutil, copy
 from collections import OrderedDict
 import pypospack.potential as potential
+import pypospack.crystal as crystal
 
 def test__import__from_pypospack_task_gulp():
     from pypospack.task.gulp import GulpSimulation
@@ -48,8 +49,23 @@ def test____init__():
             task_directory=task_directory,
             structure_filename=structure_filename,
             restart=restart)
+   
+    assert gulp_task.task_name == task_name
+    assert os.path.abspath(gulp_task.task_directory) \
+            == os.path.abspath(task_directory)
+    assert gulp_task.root_directory == os.getcwd()
+    assert os.path.exists(task_directory)
+    assert gulp_task.structure_filename == structure_filename
+    assert isinstance(
+            gulp_task.structure,
+            crystal.SimulationCell)
+    assert gulp_task.gulp_input_filename == 'gulp.in'
+    assert gulp_task.gulp_output_filename == 'gulp.out'
+    assert gulp_task.gulp_bin == os.environ['GULP_BIN']
+    assert gulp_task.status == 'INIT'
 
-    shutil.rmtree(task_directory)
+    if os.path.exists(task_directory):
+        shutil.rmtree(task_directory)
 
 def test__configure_potential():
     task_name = 'gulp_test'
@@ -69,8 +85,11 @@ def test__configure_potential():
             structure_filename=structure_filename,
             restart=restart)
 
+    assert 'potential' in configuration
+    assert 'potential_type' in configuration['potential']
+    
     gulp_task.configure_potential(
-            configuration['potential'])
+            configuration=configuration)
 
     symbols = configuration['potential']['symbols']
     parameter_names = configuration['potential']['parameter_names']
@@ -85,7 +104,41 @@ def test__configure_potential():
     for pn,pv in gulp_task.parameters.items():
         assert pv is None
 
-def test__configure():
+def test__write_gulp_input_file():
+    task_name = 'gulp_test'
+    task_directory = 'gulp_test_directory'
+    structure_filename = os.path.join(
+            'test_GulpSimulation',
+            'MgO_NaCl_prim.gga.relax.vasp')
+    restart=False
+    configuration=copy.deepcopy(configuration_MgO)
+    if os.path.exists(task_directory):
+        shutil.rmtree(task_directory)
+
+    from pypospack.task.gulp import GulpSimulation
+    gulp_task = GulpSimulation(
+            task_name=task_name,
+            task_directory=task_directory,
+            structure_filename=structure_filename,
+            restart=restart)
+    gulp_task.configuration = configuration
+    gulp_task.read_structure_file()
+    gulp_task.configure_potential()
+
+    symbols = configuration['potential']['symbols']
+    parameter_names = configuration['potential']['parameter_names']
+    
+    assert type(gulp_task.potential) is potential.BuckinghamPotential
+    assert gulp_task.potential.symbols == symbols
+    assert gulp_task.potential.parameter_names == parameter_names
+
+    assert isinstance(gulp_task.parameters,dict)
+    for pn in parameter_names:
+        assert pn in gulp_task.parameters
+    for pn,pv in gulp_task.parameters.items():
+        assert pv is None
+
+def test__on_init():
     task_name = 'gulp_test'
     task_directory = 'gulp_test_directory'
     structure_filename = os.path.join(
@@ -95,6 +148,7 @@ def test__configure():
     
     configuration=copy.deepcopy(configuration_MgO)
 
+    assert 'parameters' in configuration
     if os.path.exists(task_directory):
         shutil.rmtree(task_directory)
 
@@ -114,7 +168,7 @@ def test__configure():
     assert gulp_task.status == 'INIT'
 
     #<--- code being tested
-    gulp_task.configure(configuration=configuration)
+    gulp_task.on_init(configuration=configuration)
     
     #<------ test attribute.potential
     assert type(gulp_task.potential) is potential.BuckinghamPotential
@@ -129,42 +183,119 @@ def test__configure():
         assert pv is not None
    
     #<----- check that status update is correct
-    gulp_task.update_status()
+    assert gulp_task.conditions_CONFIG['is_structure_configured'] is True
+    assert gulp_task.conditions_CONFIG['is_potential_configured'] is True
+    assert all([v for k,v in gulp_task.conditions_INIT.items()]) 
+    assert all([v for k,v in gulp_task.conditions_CONFIG.items()]) 
+    assert gulp_task.all_conditions_INIT
+    assert gulp_task.all_conditions_CONFIG
+    assert gulp_task.all_conditions_INIT and gulp_task.all_conditions_CONFIG
     assert gulp_task.status == 'CONFIG'
 
-def test__write_gulp_input_file():
+def test__on_config():
     task_name = 'gulp_test'
     task_directory = 'gulp_test_directory'
     structure_filename = os.path.join(
             'test_GulpSimulation',
             'MgO_NaCl_prim.gga.relax.vasp')
     restart=False
+    
     configuration=copy.deepcopy(configuration_MgO)
+
     if os.path.exists(task_directory):
         shutil.rmtree(task_directory)
 
+    #<--- code setup
     from pypospack.task.gulp import GulpSimulation
     gulp_task = GulpSimulation(
             task_name=task_name,
             task_directory=task_directory,
             structure_filename=structure_filename,
             restart=restart)
-
-    gulp_task.read_structure_file()
-    gulp_task.configure_potential(configuration['potential'])
-
-    symbols = configuration['potential']['symbols']
-    parameter_names = configuration['potential']['parameter_names']
+    gulp_task.configuration = copy.deepcopy(configuration)
+    gulp_task.on_init()
     
+    #<--- precondition test
+    assert gulp_task.status == 'CONFIG'
+
+    #<--- code being tested
+    gulp_task.on_config()
+    
+    #<--- expected values
+    symbols=configuration['potential']['symbols']
+    parameter_names=configuration['potential']['parameter_names']
+    
+    #<------ test attribute.potential
     assert type(gulp_task.potential) is potential.BuckinghamPotential
     assert gulp_task.potential.symbols == symbols
     assert gulp_task.potential.parameter_names == parameter_names
 
+    #<------ test attribute.symbol
     assert isinstance(gulp_task.parameters,dict)
     for pn in parameter_names:
         assert pn in gulp_task.parameters
     for pn,pv in gulp_task.parameters.items():
-        assert pv is None
+        assert pv is not None
+   
+    #<----- check that status update is correct
+    assert gulp_task.conditions_READY['input_file_is_written']
+    assert gulp_task.all_conditions_INIT
+    assert gulp_task.all_conditions_CONFIG
+    assert gulp_task.all_conditions_READY
+    assert gulp_task.status == 'READY'
+
+def test__on_ready():
+    task_name = 'gulp_test'
+    task_directory = 'gulp_test_directory'
+    structure_filename = os.path.join(
+            'test_GulpSimulation',
+            'MgO_NaCl_prim.gga.relax.vasp')
+    restart=False
+    
+    configuration=copy.deepcopy(configuration_MgO)
+
+    if os.path.exists(task_directory):
+        shutil.rmtree(task_directory)
+
+    #<--- code setup
+    from pypospack.task.gulp import GulpSimulation
+    gulp_task = GulpSimulation(
+            task_name=task_name,
+            task_directory=task_directory,
+            structure_filename=structure_filename,
+            restart=restart)
+    gulp_task.configuration = copy.deepcopy(configuration)
+    gulp_task.on_init()
+    gulp_task.on_config()
+    #<--- precondition test
+    assert gulp_task.status == 'READY'
+
+    #<--- code being tested
+    gulp_task.on_ready()
+    
+    #<--- expected values
+    symbols=configuration['potential']['symbols']
+    parameter_names=configuration['potential']['parameter_names']
+    
+    #<------ test attribute.potential
+    assert type(gulp_task.potential) is potential.BuckinghamPotential
+    assert gulp_task.potential.symbols == symbols
+    assert gulp_task.potential.parameter_names == parameter_names
+
+    #<------ test attribute.symbol
+    assert isinstance(gulp_task.parameters,dict)
+    for pn in parameter_names:
+        assert pn in gulp_task.parameters
+    for pn,pv in gulp_task.parameters.items():
+        assert pv is not None
+   
+    #<----- check that status update is correct
+    assert gulp_task.conditions_READY['input_file_is_written']
+    assert gulp_task.all_conditions_INIT
+    assert gulp_task.all_conditions_CONFIG
+    assert gulp_task.all_conditions_READY
+    assert gulp_task.all_conditions_POST
+    assert gulp_task.status == 'POST'
 if __name__ == '__main__':
     #### TEST INITIALIZE ####
     from pypospack.task.gulp import GulpSimulation
